Split Users::PersonalKeysController into two:
- /manage/personal_key stays with Users:: controller
- /account/personal_key goes to a new Accounts:: controller
