+++ app/controllers/users/personal_keys_controller.rb
@@ -1,4 +1,5 @@
 module Users
+  # Handles updating a user's personal key if it used for 2FA
   class PersonalKeysController < ApplicationController
     include PersonalKeyConcern
     include SecureHeadersConcern
@@ -23,15 +24,6 @@ module Users
       redirect_to next_step
     end
 
-    def create
-      user_session[:personal_key] = create_new_code
-      analytics.track_event(Analytics::PROFILE_PERSONAL_KEY_CREATE)
-      create_user_event(:new_personal_key)
-      result = send_new_personal_key_notifications
-      analytics.track_event(Analytics::PROFILE_PERSONAL_KEY_CREATE_NOTIFICATIONS, result.to_h)
-      redirect_to manage_personal_key_url
-    end
-
     private
 
     def next_step
@@ -46,34 +38,6 @@ module Users
 
     def user_has_not_visited_any_sp_yet?
       current_user.identities.pluck(:last_authenticated_at).compact.empty?
-    end
-
-    # @return [FormResponse]
-    def send_new_personal_key_notifications
-      emails = current_user.confirmed_email_addresses.map do |email_address|
-        UserMailer.personal_key_regenerated(current_user, email_address.email).deliver_now_or_later
-      end
-
-      telephony_responses = MfaContext.new(current_user).
-                            phone_configurations.map do |phone_configuration|
-        phone = phone_configuration.phone
-        Telephony.send_personal_key_regeneration_notice(
-          to: phone,
-          country_code: Phonelib.parse(phone).country,
-        )
-      end
-
-      form_response(emails: emails, telephony_responses: telephony_responses)
-    end
-
-    def form_response(emails:, telephony_responses:)
-      FormResponse.new(
-        success: true,
-        extra: {
-          emails: emails.count,
-          sms_message_ids: telephony_responses.map { |resp| resp.to_h[:message_id] },
-        },
-      )
     end
   end
 end
